pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "absconda"
version = "0.1.0"
description = "Generate Dockerfiles and container build workflows from Conda environments"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "John Reeves" }]
license = { text = "MIT" }
dependencies = [
    "click>=8.1,<8.2",
    "typer[all]>=0.12,<0.13",
    "rich>=13.7,<14",
    "pydantic>=2.7,<3",
    "pyyaml>=6.0.1,<7",
    "jinja2>=3.1,<4",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9",
    "pytest-cov>=5,<6",
    "ruff>=0.6",
]

[project.scripts]
absconda = "absconda.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"absconda" = ["_templates/**/*.j2"]

[tool.pytest.ini_options]
addopts = "-ra"
pythonpath = ["src"]

[tool.ruff]
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "B"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
